requirements.txt
# Google Generative AI (Gemini)
google-genai

# TTS (Style-Bert-Vits2)
style-bert-vits2
torch>=2.0.0
torchaudio>=2.0.0

# 動画処理
moviepy==1.0.3
imageio-ffmpeg>=0.4.8

# 音声処理
pydub>=0.25.1
pysoundfile

# ユーティリティ
Pillow>=9.0.0
numpy>=1.26.0,<2.0.0
requests>=2.28.0

# その他
scipy>=1.10.0


# Marp CLI は別途インストールが必要
# npm install -g @marp-team/marp-cli

# FFmpegも必要です (MoviePy用)
# macOS: brew install ffmpeg
# Ubuntu: sudo apt install ffmpeg
# Windows: https://ffmpeg.org/download.html